U1
Validation

+++ js/login.js
@@ -61,11 +61,25 @@ var registrationApp = new Vue({
             if (this.password != "" && this.confirm_password != "" && this.password != this.confirm_password) {
                 console.log('PASSWORD NOT MATCH');
             }
+            if (!this.password.match(/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)[a-zA-Z\d]{8,}$/)) {
+                console.log('Regex Failed');
+            }
         },
         submit: function() {
-            if (this.password != "" && this.confirm_password != "" && this.password != this.confirm_password) {
-                console.log('PASSWORD NOT MATCH');
-                return alert('Password not Match')
+
+            var regexEmail = /^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$/,
+                regexPincode = /^[ABCEGHJKLMNPRSTVXYabceghjklmnprstvxy]{1}\d{1}[A-Za-z]{1}\d{1}[A-Za-z]{1}\d{1}$/, //Valid to Canadian Pincode Only
+                regexPassword = /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)[a-zA-Z\d]{8,}$/
+
+            if (this.password == "" || this.confirm_password == "" || this.password != this.confirm_password || !this.password.match(regexPassword)) {
+                return toastr.error('Password Invalid. Minimum eight characters, at least one uppercase letter, one lowercase letter and one number', 'Error');
+            }
+            if (!this.email.match(regexEmail)) {
+                return toastr.error('Email Invalid!', 'Error');
+            }
+
+            if (this.pincode == "" || !this.pincode.match(regexPincode)) {
+                return toastr.error('Pincode Invalid!', 'Error');
             }
             localStorage.setItem('email', this.email)
             localStorage.setItem('password', this.password)
